requirements.txt
openai>=1.12.0
pydantic>=2.6.0
python-dotenv>=1.0.0
rich>=13.7.0
tenacity>=8.2.0
tiktoken>=0.6.0
python-docx>=1.1.0
sympy>=1.13.0
pyyaml>=6.0
jieba>=0.42.1
openpyxl>=3.1.0
